Moved debug modal to a small template
