+++ lib/zones/debug/info.html
@@ -1,0 +1,5 @@
+<div class="ssr-debug-info">
+  <h3>{{title}}</h3>
+
+  <div>{{stack}}</div>
+</div>
